requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
#awscli
flake8
python-dotenv>=0.5.1
numpy
ipykernel
matplotlib
tensorflow[and-cuda]
opencv-python-headless
scikit-learn
mcap-ros2-support